fix10

+++ protected/controllers/TeacherController.php
@@ -656,7 +656,7 @@ class TeacherController extends CController {
                 return;
             }
         }
-        if ($_FILES["file"]["type"] == "video/mp4" || $_FILES["file"]["type"] == "application/octet-stream" && substr($_FILES["file"]["name"], strrpos($_FILES["file"]["name"], '.') + 1) != "rm" &&substr($_FILES["file"]["name"], strrpos($_FILES["file"]["name"], '.') + 1) != "rm") {
+        if ($_FILES["file"]["type"] == "video/mp4" || $_FILES["file"]["type"] == "application/octet-stream" && substr($_FILES["file"]["name"], strrpos($_FILES["file"]["name"], '.') + 1) != "rm" &&substr($_FILES["file"]["name"], strrpos($_FILES["file"]["name"], '.') + 1) != "RM") {
             if ($_FILES["file"]["error"] > 0) {
                 $result = "Return Code: " . $_FILES["file"]["error"];
             } else {
@@ -1555,7 +1555,28 @@ class TeacherController extends CController {
     public function actionDeleteLook() {
         $exerciseID = $_GET['exerciseID'];
         $thisLook = new LookType();
-        $deleteResult = $thisLook->deleteAll("exerciseID = '$exerciseID'");
+        $examLst=Array();
+        $examLst=Exam::model()->findAll();
+        $flag=0;
+        $deleteResult="";
+        $tip="";
+        foreach($examLst as $exam){
+            $examExerLst=ExamExercise::model()->getExamExerByType($exam['examID'],'look');
+            if($examExerLst!=NULL){
+                foreach($examExerLst as $examExer){
+                    if($exerciseID==$examExer['exerciseID']){
+                        $flag=1;
+                        $tip="此题目已经被占用!!!";
+                        break;
+                    }
+                }
+            }
+        }
+        if($flag==0){
+            $deleteResult = $thisLook->deleteAll("exerciseID = '$exerciseID'");
+            $tip="此题目删除成功!!!";
+        }
+        
 
         if (Yii::app()->session['lastUrl'] == "lookLst") {
             $result = LookType::model()->getLookLst("", "");
@@ -1566,7 +1587,8 @@ class TeacherController extends CController {
                 'lookLst' => $lookLst,
                 'pages' => $pages,
                 'teacher' => Teacher::model()->findall(),
-                'deleteResult' => $deleteResult
+                'deleteResult' => $deleteResult,
+                'tip'=>$tip
             ));
         } else {
             $type = Yii::app()->session['searchLookType'];
@@ -1597,7 +1619,8 @@ class TeacherController extends CController {
                 'pages' => $pages,
                 'teacher' => TbClass::model()->teaInClass(),
                 'deleteResult' => $deleteResult,
-                'searchKey' => $searchKey
+                'searchKey' => $searchKey,
+                'tip'=>$tip
             ));
         }
     }
@@ -1878,7 +1901,28 @@ class TeacherController extends CController {
     public function actionDeleteKey() {
         $exerciseID = $_GET['exerciseID'];
         $thisKey = new KeyType();
-        $deleteResult = $thisKey->deleteAll("exerciseID = '$exerciseID'");
+        $examLst=Array();
+        $examLst=Exam::model()->findAll();
+        $flag=0;
+        $deleteResult="";
+        $tip="";
+        foreach($examLst as $exam){
+            $examExerLst=ExamExercise::model()->getExamExerByType($exam['examID'],'key');
+            if($examExerLst!=NULL){
+                foreach($examExerLst as $examExer){
+                    if($exerciseID==$examExer['exerciseID']){
+                        $flag=1;
+                        $tip="此题目已经被占用!!!";
+                        break;
+                    }
+                }
+            }
+        }
+        if($flag==0){
+            $deleteResult = $thisKey->deleteAll("exerciseID = '$exerciseID'");
+            $tip="此题目删除成功!!!";
+        }
+        
 
         if (Yii::app()->session['lastUrl'] == "KeyLst") {
             $result = KeyType::model()->getKeyLst("", "");
@@ -1889,7 +1933,8 @@ class TeacherController extends CController {
                 'keyLst' => $keyLst,
                 'pages' => $pages,
                 'teacher' => Teacher::model()->findall(),
-                'deleteResult' => $deleteResult
+                'deleteResult' => $deleteResult,
+                 'tip'=>$tip
             ));
         } else {
             $type = Yii::app()->session['searchKeyType'];
@@ -1913,7 +1958,8 @@ class TeacherController extends CController {
                 'keyLst' => $keyLst,
                 'pages' => $pages,
                 'teacher' => TbClass::model()->teaInClass(),
-                'deleteResult' => $deleteResult
+                'deleteResult' => $deleteResult,
+                 'tip'=>$tip
             ));
         }
     }
@@ -2170,17 +2216,42 @@ class TeacherController extends CController {
         $exerciseID = $_GET['exerciseID'];
         $thisListen = new ListenType();
         $deleteListen = $thisListen->findAll("exerciseID = '$exerciseID'");
-        $deleteResult = $thisListen->deleteAll("exerciseID = '$exerciseID'");
-        $filePath = $deleteListen[0]['filePath'];
-        $fileName = $deleteListen[0]['fileName'];
-        if ($deleteResult == '1') {
-            $typename = Yii::app()->session['role_now'];
-            $userid = Yii::app()->session['userid_now'];
-            //怎么用EXER_LISTEN_URL
-            $path = 'resources/' . $filePath . iconv("UTF-8", "gb2312", $fileName);
-            if (file_exists($path))
-                unlink($path);
+        $deleteResult="";
+        $tip="";
+        if($deleteListen!=NULL){
+            $examLst=Array();
+            $examLst=Exam::model()->findAll();
+            $flag=0;
+            foreach($examLst as $exam){
+                $examExerLst=ExamExercise::model()->getExamExerByType($exam['examID'],'listen');
+                if($examExerLst!=NULL){
+                    foreach($examExerLst as $examExer){
+                        if($exerciseID==$examExer['exerciseID']){
+                            $flag=1;
+                            $tip="此题目已经被占用!!!";
+                            break;
+                        }
+                    }
+                }
+            }
+            if($flag==0){
+                $deleteResult = $thisListen->deleteAll("exerciseID = '$exerciseID'");
+                $tip="此题目删除成功!!!";
+            }
+
+
+            $filePath = $deleteListen[0]['filePath'];
+            $fileName = $deleteListen[0]['fileName'];
+            if ($deleteResult == '1') {
+                $typename = Yii::app()->session['role_now'];
+                $userid = Yii::app()->session['userid_now'];
+                //怎么用EXER_LISTEN_URL
+                $path = 'resources/' . $filePath . iconv("UTF-8", "gb2312", $fileName);
+                if (file_exists($path))
+                    unlink($path);
+            }
         }
+        
         if (Yii::app()->session['lastUrl'] == "listenLst") {
             $result = ListenType::model()->getListenLst("", "");
             $listenLst = $result['listenLst'];
@@ -2190,7 +2261,8 @@ class TeacherController extends CController {
                 'listenLst' => $listenLst,
                 'pages' => $pages,
                 'teacher' => Teacher::model()->findall(),
-                'deleteResult' => $deleteResult
+                'deleteResult' => $deleteResult,
+                'tip'=>$tip
             ));
         } else {
             $type = Yii::app()->session['searchListenType'];
@@ -2220,7 +2292,8 @@ class TeacherController extends CController {
                 'pages' => $pages,
                 'teacher' => TbClass::model()->teaInClass(),
                 'deleteResult' => $deleteResult,
-                'searchKey' => $searchKey
+                'searchKey' => $searchKey,
+                 'tip'=>$tip
             ));
         }
     }
@@ -2533,7 +2606,29 @@ class TeacherController extends CController {
     public function actionDeleteFill() {
         $exerciseID = $_GET["exerciseID"];
         $thisFill = new Filling();
-        $deleteResult = $thisFill->deleteAll("exerciseID = '$exerciseID'");
+        $examLst=Array();
+        $examLst=Exam::model()->findAll();
+        $flag=0;
+        $deleteResult="";
+        $tip="";
+        foreach($examLst as $exam){
+            $examExerLst=ExamExercise::model()->getExamExerByType($exam['examID'],'filling');
+            if($examExerLst!=NULL){
+                foreach($examExerLst as $examExer){
+                    if($exerciseID==$examExer['exerciseID']){
+                        $flag=1;
+                        $tip="此题目已经被占用!!!";
+                        break;
+                    }
+                }
+            }
+        }
+        if($flag==0){
+            $deleteResult = $thisFill->deleteAll("exerciseID = '$exerciseID'");
+            $tip="此题目删除成功!!!";
+        }
+        
+        
         if (Yii::app()->session['lastUrl'] == "searchFill") {
             $type = Yii::app()->session['searchFillType'];
             $value = Yii::app()->session['searchFillValue'];
@@ -2563,7 +2658,8 @@ class TeacherController extends CController {
                 'pages' => $pages,
                 'teacher' => Teacher::model()->findall(),
                 'deleteResult' => $deleteResult,
-                'searchKey' => $searchKey
+                'searchKey' => $searchKey,
+                'tip'=>$tip
                     )
             );
         } else {
@@ -2575,7 +2671,8 @@ class TeacherController extends CController {
                 'fillLst' => $fillLst,
                 'pages' => $pages,
                 'teacher' => Teacher::model()->findall(),
-                'deleteResult' => $deleteResult
+                'deleteResult' => $deleteResult,
+                'tip'=>$tip
             ));
         }
     }
@@ -2877,7 +2974,27 @@ class TeacherController extends CController {
     public function actionDeleteChoice() {
         $exerciseID = $_GET["exerciseID"];
         $thisChoice = new Choice();
-        $deleteResult = $thisChoice->deleteAll("exerciseID = '$exerciseID'");
+        $examLst=Array();
+        $examLst=Exam::model()->findAll();
+        $flag=0;
+        $deleteResult="";
+        $tip="";
+        foreach($examLst as $exam){
+            $examExerLst=ExamExercise::model()->getExamExerByType($exam['examID'],'choice');
+            if($examExerLst!=NULL){
+                foreach($examExerLst as $examExer){
+                    if($exerciseID==$examExer['exerciseID']){
+                        $flag=1;
+                        $tip="此题目已经被占用!!!";
+                        break;
+                    }
+                }
+            }
+        }
+        if($flag==0){
+            $deleteResult = $thisChoice->deleteAll("exerciseID = '$exerciseID'");
+            $tip="此题目删除成功!!!";
+        }
         if (Yii::app()->session['lastUrl'] == "searchChoice") {
             $type = Yii::app()->session['searchChoiceType'];
             $value = Yii::app()->session['searchChoiceValue'];
@@ -2918,6 +3035,7 @@ class TeacherController extends CController {
                 'pages' => $pages,
                 'teacher' => Teacher::model()->findall(),
                 'deleteResult' => $deleteResult,
+                'tip'=>$tip,
             ));
         }
     }
@@ -3139,7 +3257,29 @@ class TeacherController extends CController {
     public function actionDeleteQuestion() {
         $exerciseID = $_GET["exerciseID"];
         $thisQuestion = new Question();
-        $deleteResult = $thisQuestion->deleteAll("exerciseID = '$exerciseID'");
+        $examLst=Array();
+        $examLst=Exam::model()->findAll();
+        $flag=0;
+        $deleteResult="";
+        $tip="";
+        foreach($examLst as $exam){
+            $examExerLst=ExamExercise::model()->getExamExerByType($exam['examID'],'question');
+            if($examExerLst!=NULL){
+                foreach($examExerLst as $examExer){
+                    if($exerciseID==$examExer['exerciseID']){
+                        $flag=1;
+                        $tip="此题目已经被占用!!!";
+                        break;
+                    }
+                }
+            }
+        }
+        if($flag==0){
+            $deleteResult = $thisQuestion->deleteAll("exerciseID = '$exerciseID'");
+            $tip="此题目删除成功!!!";
+        }
+        
+        
         if (Yii::app()->session['lastUrl'] == "searchQuestion") {
             $type = Yii::app()->session['searchQuestionType'];
             $value = Yii::app()->session['searchQuestionValue'];
@@ -3167,7 +3307,8 @@ class TeacherController extends CController {
                 'pages' => $pages,
                 'teacher' => Teacher::model()->findall(),
                 'deleteResult' => $deleteResult,
-                'searchKey' => $searchKey
+                'searchKey' => $searchKey,
+                'tip'=>$tip
                     )
             );
         } else {
@@ -3179,7 +3320,8 @@ class TeacherController extends CController {
                 'questionLst' => $questionLst,
                 'pages' => $pages,
                 'teacher' => Teacher::model()->findall(),
-                'deleteResult' => $deleteResult
+                'deleteResult' => $deleteResult,
+                'tip'=>$tip
             ));
         }
     }
